Room.java checkin() - Unit testing
In this commit, the method checkin has been completed. This method checks whether the state is set to ready or not. If the state is not set to ready then it throws runtimeException message. If the state is ready then it will set the state to occupied. Then, the method ends. General comments were also added in the code to make it easy to understand.