Room.java checkout() - Unit Testing
In this commit, method named checkout() is created which takes booking as an argument. The method checks if the state is set to occupied or not. If the state is not set to occupied then it shows the runtimeException message. If the state is set to occupied then it removes the new booking named booking from bookings and then sets the state to  ready. Then, the method ends.