RecordServiceCTL - Integration Testing
In this commit, the method serviceDetailsEntered has been completed. This method checks whether state is service or not. If the state is not service it gives exception message and if the state is service then it calls addServiceCharge and ui displayServiceChargeMessage. Then state is changed to completed and ui state is changed to completed.