BookingCTL.creditDetailsEntered() - Integration Testing
As per Pseudo Code, this method first it checks if the state is CREDIT or not.IF it is CREDIT, it shows RunTimeException Error. Then it creates a CreditCard Object and gets the type, number and ccv from user. Later a boolean approved is initializaed that will call the CreditAuthorizer to check if the Card details entered are valid or not. If approved od returned Ture, it will get teh confirmationNumber from hotel.Book(). Later various variables are initialized to match the bookingUI.displayConfirmedMEssage parameters. IT displayes the confirmationNumber and changes the state ot COMPLETED. And even changes to bookingUI state to COMPLETED. Now if the boolean approved was false, it will display that the credit card was not approved.